add tvm stest
