+++ src/test/java/stest/tron/wallet/dailybuild/tvmnewcommand/transferfailed/TransferFailed005.java
@@ -1,4 +1,4 @@
-package stest.tron.wallet.dailybuild.TransferFailed;
+package stest.tron.wallet.dailybuild.tvmnewcommand.transferfailed;
 
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
@@ -26,7 +26,7 @@ import stest.tron.wallet.common.client.utils.Base58;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 
 @Slf4j
-public class TransferFailed006 {
+public class TransferFailed005 {
 
   private final String testNetAccountKey = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
@@ -84,7 +84,7 @@ public class TransferFailed006 {
             blockingStubFull));
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
-    String filePath = "src/test/resources/soliditycode/TransferFailed006.sol";
+    String filePath = "src/test/resources/soliditycode/TransferFailed005.sol";
     String contractName = "EnergyOfTransferFailedTest";
     HashMap retMap = PublicMethed.getBycodeAbi(filePath, contractName);
     String code = retMap.get("byteCode").toString();
@@ -99,7 +99,7 @@ public class TransferFailed006 {
     contractAddress = infoById.get().getContractAddress().toByteArray();
     Assert.assertEquals(0, infoById.get().getResultValue());
 
-    filePath = "src/test/resources/soliditycode/TransferFailed006.sol";
+    filePath = "src/test/resources/soliditycode/TransferFailed005.sol";
     contractName = "Caller";
     retMap = PublicMethed.getBycodeAbi(filePath, contractName);
     code = retMap.get("byteCode").toString();
@@ -116,7 +116,7 @@ public class TransferFailed006 {
     Assert.assertEquals(0, infoById.get().getResultValue());
   }
 
-  @Test(enabled = true, description = "TransferFailed for create")
+  @Test(enabled = true, description = "TransferFailed for function call_value")
   public void triggerContract() {
     Account info = null;
 
@@ -148,14 +148,59 @@ public class TransferFailed006 {
             .queryAccount(contractAddress1, blockingStubFull)
             .getBalance());
     long paramValue = 1000000;
-    String param = "\"" + paramValue + "\"";
 
+    // transfer trx to self`s account
+    String param = "\"" + paramValue + "\",\"" + Base58.encode58Check(contractAddress) + "\"";
     String triggerTxid = PublicMethed.triggerContract(contractAddress,
-        "testCreateTrxInsufficientBalance(uint256)", param, false, 0L,
+        "testCallTrxInsufficientBalance(uint256,address)", param, false, 0L,
+        maxFeeLimit, accountExcAddress, accountExcKey, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    Optional<TransactionInfo> infoById = PublicMethed
+        .getTransactionInfoById(triggerTxid, blockingStubFull);
+
+    Assert.assertEquals(infoById.get().getResultValue(), 1);
+    Assert.assertEquals("FAILED", infoById.get().getResult().toString());
+    Assert.assertEquals("TRANSFER_FAILED", infoById.get().getReceipt().getResult().toString());
+    Assert.assertEquals("transfer trx failed: Cannot transfer trx to yourself.",
+        infoById.get().getResMessage().toStringUtf8());
+    Assert.assertEquals(1000100L,
+        PublicMethed.queryAccount(contractAddress, blockingStubFull).getBalance());
+    Assert.assertEquals(1L,
+        PublicMethed.queryAccount(contractAddress1, blockingStubFull).getBalance());
+    Assert.assertTrue(infoById.get().getReceipt().getEnergyUsageTotal() < 10000000);
+
+    // transfer trx to unactivate account
+    ECKey ecKey2 = new ECKey(Utils.getRandom());
+    byte[] accountExcAddress2 = ecKey2.getAddress();
+    param = "\"" + paramValue + "\",\"" + Base58.encode58Check(accountExcAddress2) + "\"";
+    triggerTxid = PublicMethed.triggerContract(contractAddress,
+        "testCallTrxInsufficientBalance(uint256,address)", param, false, 0L,
+        maxFeeLimit, accountExcAddress, accountExcKey, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    infoById = PublicMethed
+        .getTransactionInfoById(triggerTxid, blockingStubFull);
+
+    Assert.assertEquals(infoById.get().getResultValue(), 1);
+    Assert.assertEquals("FAILED", infoById.get().getResult().toString());
+    Assert.assertEquals("TRANSFER_FAILED", infoById.get().getReceipt().getResult().toString());
+    Assert.assertEquals(
+        "transfer trx failed: Validate InternalTransfer error, no ToAccount. "
+            + "And not allowed to create account in smart contract.",
+        infoById.get().getResMessage().toStringUtf8());
+    Assert.assertEquals(1000100L,
+        PublicMethed.queryAccount(contractAddress, blockingStubFull).getBalance());
+    Assert.assertEquals(1L,
+        PublicMethed.queryAccount(contractAddress1, blockingStubFull).getBalance());
+    Assert.assertTrue(infoById.get().getReceipt().getEnergyUsageTotal() < 10000000);
+
+    // transfer trx to caller, value enough , function success contractResult(call_value) failed
+    param = "\"" + paramValue + "\",\"" + Base58.encode58Check(contractAddress1) + "\"";
+    triggerTxid = PublicMethed.triggerContract(contractAddress,
+        "testCallTrxInsufficientBalance(uint256,address)", param, false, 0L,
         maxFeeLimit, accountExcAddress, accountExcKey, blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
-    Optional<TransactionInfo> infoById = PublicMethed
+    infoById = PublicMethed
         .getTransactionInfoById(triggerTxid, blockingStubFull);
     logger.info(infoById.get().getReceipt().getResult() + "");
 
@@ -170,22 +215,22 @@ public class TransferFailed006 {
     logger.info("netFee:" + netFee);
     logger.info("energyUsageTotal:" + energyUsageTotal);
 
-    logger.info(
-        "contractAddress balance before: " + PublicMethed
-            .queryAccount(contractAddress, blockingStubFull)
-            .getBalance());
-    logger.info(
-        "callerAddress balance before: " + PublicMethed
-            .queryAccount(contractAddress1, blockingStubFull)
-            .getBalance());
+    int contractResult = ByteArray
+        .toInt(infoById.get().getContractResult(0).toByteArray());
+    Assert.assertEquals(1, contractResult);
+
     Assert.assertEquals(infoById.get().getResultValue(), 0);
-    Assert.assertFalse(infoById.get().getInternalTransactions(0).getRejected());
+    Assert.assertEquals(infoById.get().getResult().toString(), "SUCESS");
     Assert.assertEquals(100L,
         PublicMethed.queryAccount(contractAddress, blockingStubFull).getBalance());
+    Assert.assertEquals(1000001L,
+        PublicMethed.queryAccount(contractAddress1, blockingStubFull).getBalance());
     Assert.assertTrue(infoById.get().getReceipt().getEnergyUsageTotal() < 10000000);
 
+    // transfer trx to caller, value not enough, function success
+    // but contractResult(call_value) failed
     triggerTxid = PublicMethed.triggerContract(contractAddress,
-        "testCreateTrxInsufficientBalance(uint256)", param, false, 0L,
+        "testCallTrxInsufficientBalance(uint256,address)", param, false, 0L,
         maxFeeLimit, accountExcAddress, accountExcKey, blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
@@ -201,20 +246,15 @@ public class TransferFailed006 {
     logger.info("netFee:" + netFee);
     logger.info("energyUsageTotal:" + energyUsageTotal);
 
-    logger.info(
-        "contractAddress balance before: " + PublicMethed
-            .queryAccount(contractAddress, blockingStubFull)
-            .getBalance());
-    logger.info(
-        "callerAddress balance before: " + PublicMethed
-            .queryAccount(contractAddress1, blockingStubFull)
-            .getBalance());
-
-    Assert.assertEquals(infoById.get().getResultValue(), 1);
-    Assert.assertEquals(infoById.get().getResMessage().toStringUtf8(), "REVERT opcode executed");
+    //contractResult`s first boolean value
+    contractResult = ByteArray
+        .toInt(infoById.get().getContractResult(0).toByteArray());
+    Assert.assertEquals(0, contractResult);
+    Assert.assertEquals(infoById.get().getResultValue(), 0);
+    Assert.assertEquals(infoById.get().getResult().toString(), "SUCESS");
     Assert.assertEquals(100L,
         PublicMethed.queryAccount(contractAddress, blockingStubFull).getBalance());
-    Assert.assertEquals(1L,
+    Assert.assertEquals(1000001L,
         PublicMethed.queryAccount(contractAddress1, blockingStubFull).getBalance());
     Assert.assertTrue(infoById.get().getReceipt().getEnergyUsageTotal() < 10000000);
